Moved functions to Util. Added JSDoc.
