+++ src/pages/_document.tsx
@@ -1,6 +1,6 @@
 import {Html, Head, Main, NextScript} from 'next/document';
 
-export default function Document() {
+const Document = () => {
   return (
     <Html lang="en">
       <Head />
@@ -10,4 +10,6 @@ export default function Document() {
       </body>
     </Html>
   );
-}
+};
+
+export default Document;
