+++ src/pages/index.tsx
@@ -4,10 +4,12 @@ import {Showcase} from '@/components/Showcase';
 import {WebComponent} from '@/components/WebComponent';
 import {Brand, Dollhouse, Help, OpenBehavior, Play, Quickstart, Tour, GuidedTour, HighlightReel, MlsBehavior, Mattertags, TagNavigation, Pin, Portal, SwitchFloors, FloorPlanView, Language, Zoom, Search, Wheel, GuidedTourPan, LoopBack, Title, GuidedTourCallToAction, HighlightReelBehavior, VirtualReality} from '@/types/enums';
 
-const a = 'b';
-
-//TODO: Check out Intersection Observer API - https://www.dhiwise.com/post/implementing-next-js-lazy-loading-for-optimized-web-apps
-//TODO: Add getStaticProps() - https://blog.logrocket.com/using-next-js-with-typescript/
+/**
+ * Home component. MpViewer wrapper
+ *
+ * @returns {JSX.Element} Component
+ * @component
+ */
 const Home = () => {
   const [sdk, setSdk] = useState<MpSdk>(null);
   const configSdk: SdkConfig = {
@@ -88,7 +90,6 @@ const Home = () => {
 
   return (
     <div className="flex flex-col h-screen p-2 bg-gray-200">
-      {/* TODO: Avoid 2 instances loading */}
       <h2>Showcase SDK</h2>
       <Showcase config={configSdk} setMpSdk={setSdk} />
